requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
sqlalchemy
pydantic

# Splunk integration
splunk-sdk